# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/mogrtgage/cache": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Get all cache",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entity.CachedMortgage"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/v1.response"
                        }
                    }
                }
            }
        },
        "/mogrtgage/execute": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Execute credit",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.executeResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/v1.response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "entity.CachedMortgage": {
            "type": "object",
            "properties": {
                "aggregates": {
                    "$ref": "#/definitions/mortgage.Aggregates"
                },
                "id": {
                    "type": "integer"
                },
                "params": {
                    "$ref": "#/definitions/mortgage.Params"
                },
                "program": {
                    "$ref": "#/definitions/mortgage.Program"
                }
            }
        },
        "entity.Mortgage": {
            "type": "object",
            "properties": {
                "aggregates": {
                    "$ref": "#/definitions/mortgage.Aggregates"
                },
                "params": {
                    "$ref": "#/definitions/mortgage.Params"
                },
                "program": {
                    "$ref": "#/definitions/mortgage.Program"
                }
            }
        },
        "mortgage.Aggregates": {
            "type": "object",
            "properties": {
                "last_payment_date": {
                    "type": "string"
                },
                "loan_sum": {
                    "type": "integer"
                },
                "monthly_payment": {
                    "type": "integer"
                },
                "overpayment": {
                    "type": "integer"
                },
                "rate": {
                    "type": "integer"
                }
            }
        },
        "mortgage.Params": {
            "type": "object",
            "properties": {
                "initial_payment": {
                    "type": "integer"
                },
                "months": {
                    "type": "integer"
                },
                "object_cost": {
                    "type": "integer"
                }
            }
        },
        "mortgage.Program": {
            "type": "object",
            "properties": {
                "base": {
                    "type": "boolean"
                },
                "military": {
                    "type": "boolean"
                },
                "salary": {
                    "type": "boolean"
                }
            }
        },
        "v1.executeResponse": {
            "type": "object",
            "properties": {
                "result": {
                    "$ref": "#/definitions/entity.Mortgage"
                }
            }
        },
        "v1.response": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "message"
                }
            }
        }
    }
}